+++ src/num_traits/scale.rs
@@ -5,7 +5,7 @@ pub trait Scale<T> {
 
 #[macro_export]
 macro_rules! impl_scale {
-    ($t:ident < $g:ident >) => {
+    ($t:ident < $g:ident > $([$(<$a:ident $(, $p:tt)?>),*])?) => {
         impl<$g: $crate::Scale<O>, O> $crate::Scale<$t<O>> for $t<$g> {
             fn scale(self) -> $t<O> {
                 self.map($g::scale)