+++ src/rgba.rs
@@ -197,6 +197,10 @@ impl<T> MapExt for Rgba<T> {
     fn map<R>(self, f: impl FnMut(Self::Val) -> R) -> Self::This<R> {
         Rgba(self.0.map(f))
     }
+
+    fn mutate(&mut self, f: impl FnMut(&mut Self::Val)) {
+        self.0.mutate(f);
+    }
 }
 
 impl_traits!(Rgba<T> => MapTraits);