+++ src/vec3.rs
@@ -9,8 +9,9 @@ use std::{
 };
 
 use crate::{
-    Checked, Float, Goniometric, IntoFloat, Isqrt, LargeType, MapExt,
-    NormalLimits, Scale, Sqrt, Vec2, Zero, impl_traits,
+    Array, Checked, CompArithm, Float, Goniometric, IntoFloat, Isqrt,
+    LargeType, MapExt, NormalLimits, Scale, Sqrt, Tuple, Vec2, Zero,
+    impl_traits,
 };
 
 /// Represents three dimensional vector. Can be also use as color or any
@@ -1092,25 +1093,6 @@ impl<'a, T> IntoIterator for &'a mut Vec3<T> {
     }
 }
 
-impl<T: Neg> Neg for Vec3<T> {
-    type Output = Vec3<T::Output>;
-
-    fn neg(self) -> Self::Output {
-        self.map(|x| -x)
-    }
-}
-
-impl<T> Not for Vec3<T>
-where
-    T: Not,
-{
-    type Output = Vec3<T::Output>;
-
-    fn not(self) -> Self::Output {
-        self.map(|a| !a)
-    }
-}
-
 impl<T: PartialOrd> PartialOrd for Vec3<T> {
     fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
         match (
@@ -1136,126 +1118,19 @@ impl<T: PartialOrd> PartialOrd for Vec3<T> {
     }
 }
 
-impl_traits!(Vec3<T> => VecTraits);
+impl<T> MapExt for Vec3<T> {
+    type Val = T;
+    type This<R> = Vec3<R>;
 
-macro_rules! op_single {
-    ($op:ident, $fn:ident) => {
-        impl<Left, Right> $op<Right> for Vec3<Left>
-        where
-            Left: $op<Right>,
-            Right: Copy,
-        {
-            type Output = Vec3<Left::Output>;
+    fn map<R>(self, mut f: impl FnMut(Self::Val) -> R) -> Self::This<R> {
+        Vec3::new(f(self.x), f(self.y), f(self.z))
+    }
 
-            fn $fn(self, rhs: Right) -> Self::Output {
-                self.map(|x| x.$fn(rhs))
-            }
-        }
-    };
+    fn mutate(&mut self, mut f: impl FnMut(&mut Self::Val)) {
+        f(&mut self.x);
+        f(&mut self.y);
+        f(&mut self.z);
+    }
 }
 
-macro_rules! op_assign_single {
-    ($op:ident, $fn:ident) => {
-        impl<Left, Right> $op<Right> for Vec3<Left>
-        where
-            Left: $op<Right>,
-            Right: Copy,
-        {
-            fn $fn(&mut self, rhs: Right) {
-                self.x.$fn(rhs);
-                self.y.$fn(rhs);
-                self.z.$fn(rhs);
-            }
-        }
-    };
-}
-
-macro_rules! op_triple {
-    ($op:ident, $fn:ident) => {
-        impl<Left, Right> $op<Vec3<Right>> for Vec3<Left>
-        where
-            Left: $op<Right>,
-        {
-            type Output = Vec3<Left::Output>;
-
-            fn $fn(self, rhs: Vec3<Right>) -> Self::Output {
-                (self.x.$fn(rhs.x), self.y.$fn(rhs.y), self.z.$fn(rhs.z))
-                    .into()
-            }
-        }
-
-        impl<Left, Right> $op<(Right, Right, Right)> for Vec3<Left>
-        where
-            Left: $op<Right>,
-        {
-            type Output = Vec3<Left::Output>;
-
-            fn $fn(self, (x, y, z): (Right, Right, Right)) -> Self::Output {
-                (self.x.$fn(x), self.y.$fn(y), self.z.$fn(z)).into()
-            }
-        }
-
-        impl<Left, Right> $op<[Right; 3]> for Vec3<Left>
-        where
-            Left: $op<Right>,
-        {
-            type Output = Vec3<Left::Output>;
-
-            fn $fn(self, [x, y, z]: [Right; 3]) -> Self::Output {
-                (self.x.$fn(x), self.y.$fn(y), self.z.$fn(z)).into()
-            }
-        }
-    };
-}
-
-macro_rules! op_assign_triple {
-    ($op:ident, $fn:ident) => {
-        impl<Left, Right> $op<Vec3<Right>> for Vec3<Left>
-        where
-            Left: $op<Right>,
-        {
-            fn $fn(&mut self, rhs: Vec3<Right>) {
-                self.x.$fn(rhs.x);
-                self.y.$fn(rhs.y);
-                self.z.$fn(rhs.z);
-            }
-        }
-
-        impl<Left, Right> $op<(Right, Right, Right)> for Vec3<Left>
-        where
-            Left: $op<Right>,
-        {
-            fn $fn(&mut self, (x, y, z): (Right, Right, Right)) {
-                self.x.$fn(x);
-                self.y.$fn(y);
-                self.z.$fn(z);
-            }
-        }
-
-        impl<Left, Right> $op<[Right; 3]> for Vec3<Left>
-        where
-            Left: $op<Right>,
-        {
-            fn $fn(&mut self, [x, y, z]: [Right; 3]) {
-                self.x.$fn(x);
-                self.y.$fn(y);
-                self.z.$fn(z);
-            }
-        }
-    };
-}
-
-op_single!(Mul, mul);
-op_assign_single!(MulAssign, mul_assign);
-
-op_single!(Div, div);
-op_assign_single!(DivAssign, div_assign);
-
-op_single!(Rem, rem);
-op_assign_single!(RemAssign, rem_assign);
-
-op_triple!(Add, add);
-op_assign_triple!(AddAssign, add_assign);
-
-op_triple!(Sub, sub);
-op_assign_triple!(SubAssign, sub_assign);
+impl_traits!(Vec3<T> [<Array, 3>, <Tuple, (x, y, z)>] => VecTraits);